Datasets refactoring
Extracted glue_profiling_handler
