+++ backend/dataall/aws/handlers/glue.py
@@ -6,8 +6,6 @@ from .service_handlers import Worker
 from .sts import SessionHelper
 from ... import db
 from ...db import models
-from dataall.modules.datasets.db.models import DatasetProfilingRun
-from dataall.modules.datasets.services.dataset_profiling_service import DatasetProfilingService
 
 log = logging.getLogger('aws:glue')
 
@@ -523,90 +521,6 @@ class Glue:
                 log.warning(f'Could not retrieve pipeline runs , {str(e)}')
                 return []
             return response['JobRuns']
-
-    @staticmethod
-    @Worker.handler('glue.job.start_profiling_run')
-    def start_profiling_run(engine, task: models.Task):
-        with engine.scoped_session() as session:
-            profiling: DatasetProfilingRun = (
-                DatasetProfilingService.get_profiling_run(
-                    session, profilingRunUri=task.targetUri
-                )
-            )
-            dataset: models.Dataset = session.query(models.Dataset).get(
-                profiling.datasetUri
-            )
-            run = Glue.run_job(
-                **{
-                    'accountid': dataset.AwsAccountId,
-                    'name': dataset.GlueProfilingJobName,
-                    'region': dataset.region,
-                    'arguments': (
-                        {'--table': profiling.GlueTableName}
-                        if profiling.GlueTableName
-                        else {}
-                    ),
-                }
-            )
-            DatasetProfilingService.update_run(
-                session,
-                profilingRunUri=profiling.profilingRunUri,
-                GlueJobRunId=run['JobRunId'],
-            )
-            return run
-
-    @staticmethod
-    def run_job(**data):
-        accountid = data['accountid']
-        name = data['name']
-        try:
-            session = SessionHelper.remote_session(accountid=accountid)
-            client = session.client('glue', region_name=data.get('region', 'eu-west-1'))
-            response = client.start_job_run(
-                JobName=name, Arguments=data.get('arguments', {})
-            )
-            return response
-        except ClientError as e:
-            log.error(f'Failed to start profiling job {name} due to: {e}')
-            raise e
-
-    @staticmethod
-    @Worker.handler('glue.job.profiling_run_status')
-    def get_profiling_run(engine, task: models.Task):
-        with engine.scoped_session() as session:
-            profiling: DatasetProfilingRun = (
-                DatasetProfilingService.get_profiling_run(
-                    session, profilingRunUri=task.targetUri
-                )
-            )
-            dataset: models.Dataset = session.query(models.Dataset).get(
-                profiling.datasetUri
-            )
-            glue_run = Glue.get_job_run(
-                **{
-                    'accountid': dataset.AwsAccountId,
-                    'name': dataset.GlueProfilingJobName,
-                    'region': dataset.region,
-                    'run_id': profiling.GlueJobRunId,
-                }
-            )
-            profiling.status = glue_run['JobRun']['JobRunState']
-            session.commit()
-            return profiling.status
-
-    @staticmethod
-    def get_job_run(**data):
-        accountid = data['accountid']
-        name = data['name']
-        run_id = data['run_id']
-        try:
-            session = SessionHelper.remote_session(accountid=accountid)
-            client = session.client('glue', region_name=data.get('region', 'eu-west-1'))
-            response = client.get_job_run(JobName=name, RunId=run_id)
-            return response
-        except ClientError as e:
-            log.error(f'Failed to get job run {run_id} due to: {e}')
-            raise e
 
     @staticmethod
     def grant_principals_all_table_permissions(
